+++ R/split_vector.R
@@ -11,6 +11,7 @@
 #' @return Returns a list of vectors. 
 #' @author Matthew Flickinger and Tyler Rinker <tyler.rinker@@gmail.com>.
 #' @references \url{http://stackoverflow.com/a/24319217/1000343} 
+#' @seealso \code{\link[qdapTools]{loc_split}}
 #' @export
 #' @examples
 #' set.seed(15)